Final refactor pass
